Feature/controler layer

+++ app/dtos/rideDto.go
@@ -1,6 +1,6 @@
 package dtos
 
 type RideDto struct {
-	IdUser    int64 `json:"idUser"`
-	IdVehicle int64 `json:"idVehicle"`
+	IdUser    int `json:"idUser"`
+	IdVehicle int `json:"idVehicle"`
 }
